+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+- Added directive `export 'package:template_block/template_block.dart'`
+- Added setter `body: Generator` to class `MethodGenerator`
+
 ## 0.0.2
 
 - Implemented directive generators (`export`, `import`, `library`, `part`, `part of`)